Refactor notification icon resolution logic and enhance icon path handling
